requirements.txt
numpy>=1.20.0
opencv-python>=4.5.0
scipy>=1.7.0
scikit-learn>=1.0.0
pyyaml>=5.4.0
dataclasses>=0.6; python_version<"3.7"